seperate forest generation
